Update main.py
Cleaned some things up and now cards are discarded to decide your hand! Cards are appended to the crib and are highlighted when selected in the discarding phase. The 'go' button is used as a 'I am done picking my discarded cards' button. So versatile. Next up is counting the hand itself after pegging, having a cut-card, and then of course playing multiple rounds in a row