Update main.py
Cards are now not immediately shown and need to be revealed for the player's turn. Allows players to play on one device and not see each others' hands. Also, rounds can be played now! Everything resets for full games to potentially be played. Still some bugs to work out with new revealing implementation but nearly playable!