Update main.py
Things are broken while I rearrange everything to work with the new implementation. Everything will look much prettier and be much more efficient once it's over with. To make more progress, things need to be readable and easy to change, which this will enable. A little bit of a headache to rework everything but it'll be worth the effort long-term.